#13818 - Move cron expression validation into the api module so the user interface can share it
