+++ sormas-backend/src/main/java/de/symeda/sormas/backend/common/cron/CronExpressionParser.java
@@ -15,98 +15,27 @@
 
 package de.symeda.sormas.backend.common.cron;
 
-import java.util.regex.Pattern;
-
 import javax.ejb.ScheduleExpression;
 
+import de.symeda.sormas.api.systemconfiguration.CronExpressionValidator;
+
 public final class CronExpressionParser {
-
-	public static final String FIELD_PATTERN = "(\\*|\\*/\\d{1,2}|\\d{1,2}(-\\d{1,2})?(/\\d{1,2})?(,\\d{1,2}(-\\d{1,2})?)*)";
-	public static final String VALUE_PATTERN = "(|" + FIELD_PATTERN + "( " + FIELD_PATTERN + "){5})";
-
-	private static final Pattern STRUCTURE = Pattern.compile(VALUE_PATTERN);
-
-	private static final int SECOND = 0;
-	private static final int MINUTE = 1;
-	private static final int HOUR = 2;
-	private static final int DAY_OF_MONTH = 3;
-	private static final int MONTH = 4;
-	private static final int DAY_OF_WEEK = 5;
-
-	private static final int[] LOWER_BOUNDS = {
-		0, 0, 0, 1, 1, 0 };
-	private static final int[] UPPER_BOUNDS = {
-		59, 59, 23, 31, 12, 7 };
 
 	private CronExpressionParser() {
 	}
 
-	public static boolean isDisabled(String expression) {
-		return expression == null || expression.trim().isEmpty();
-	}
-
-	public static boolean isValid(String expression) {
-		return isDisabled(expression) || toScheduleExpression(expression) != null;
-	}
-
 	public static ScheduleExpression parse(String expression) {
 
-		ScheduleExpression parsed = toScheduleExpression(expression);
-		if (parsed == null) {
+		if (CronExpressionValidator.isDisabled(expression) || !CronExpressionValidator.isValid(expression)) {
 			throw new IllegalArgumentException("Not a valid cron expression: [" + expression + "]");
-		}
-		return parsed;
-	}
-
-	private static ScheduleExpression toScheduleExpression(String expression) {
-
-		if (expression == null || !STRUCTURE.matcher(expression).matches() || expression.isEmpty()) {
-			return null;
 		}
 
 		String[] fields = expression.split(" ");
-		for (int field = 0; field < fields.length; field++) {
-			if (!isWithinBounds(field, fields[field], LOWER_BOUNDS[field], UPPER_BOUNDS[field])) {
-				return null;
-			}
-		}
-
-		return new ScheduleExpression().second(fields[SECOND])
-			.minute(fields[MINUTE])
-			.hour(fields[HOUR])
-			.dayOfMonth(fields[DAY_OF_MONTH])
-			.month(fields[MONTH])
-			.dayOfWeek(fields[DAY_OF_WEEK]);
-	}
-
-	private static boolean isWithinBounds(int fieldIndex, String field, int lowerBound, int upperBound) {
-
-		for (String listEntry : field.split(",")) {
-			String[] baseAndIncrement = listEntry.split("/");
-			String withoutIncrement = baseAndIncrement[0];
-			if (baseAndIncrement.length > 1) {
-				if (!allowsIncrement(fieldIndex)) {
-					return false;
-				}
-				int increment = Integer.parseInt(baseAndIncrement[1]);
-				if (increment < 1 || increment > upperBound) {
-					return false;
-				}
-			}
-			if ("*".equals(withoutIncrement)) {
-				continue;
-			}
-			for (String bound : withoutIncrement.split("-")) {
-				int value = Integer.parseInt(bound);
-				if (value < lowerBound || value > upperBound) {
-					return false;
-				}
-			}
-		}
-		return true;
-	}
-
-	private static boolean allowsIncrement(int fieldIndex) {
-		return fieldIndex == SECOND || fieldIndex == MINUTE || fieldIndex == HOUR;
+		return new ScheduleExpression().second(fields[CronExpressionValidator.SECOND])
+			.minute(fields[CronExpressionValidator.MINUTE])
+			.hour(fields[CronExpressionValidator.HOUR])
+			.dayOfMonth(fields[CronExpressionValidator.DAY_OF_MONTH])
+			.month(fields[CronExpressionValidator.MONTH])
+			.dayOfWeek(fields[CronExpressionValidator.DAY_OF_WEEK]);
 	}
 }
